Move function schema code to new schema package
