+++ dgraph/dgraph.go
@@ -8,11 +8,6 @@ import (
 	"context"
 	"fmt"
 	"hmruntime/utils"
-	"log"
-
-	"github.com/dgraph-io/gqlparser/ast"
-	"github.com/dgraph-io/gqlparser/parser"
-	"github.com/dgraph-io/gqlparser/validator"
 )
 
 var DgraphUrl *string
@@ -83,105 +78,6 @@ func GetGQLSchema(ctx context.Context) (string, error) {
 	}
 
 	return data.Node[0].Schema, nil
-}
-
-type FunctionInfo struct {
-	PluginName string
-	Schema     functionSchema
-}
-
-type functionSchema struct {
-	ObjectDef *ast.Definition
-	FieldDef  *ast.FieldDefinition
-}
-
-func (info FunctionInfo) FunctionName() string {
-	return info.Schema.FunctionName()
-}
-
-func (schema functionSchema) Resolver() string {
-	return schema.ObjectDef.Name + "." + schema.FieldDef.Name
-}
-
-func (schema functionSchema) FunctionName() string {
-	f := schema.FieldDef
-
-	// If @hm_function(name: "name") is specified, use that.
-	d := f.Directives.ForName("hm_function")
-	if d != nil {
-		a := d.Arguments.ForName("name")
-		if a != nil && a.Value != nil {
-			return a.Value.Raw
-		}
-	}
-
-	// No @hm_function directive, or no name argument. Just use the field name.
-	return f.Name
-}
-
-func (schema functionSchema) FunctionArgs() ast.ArgumentDefinitionList {
-	f := schema.FieldDef
-
-	// If @hm_function(args: ["arg1", "arg2"]) is specified, use that.
-	// The arguments must correspond to field names on the same parent object.
-	// The types will be ascertained from the corresponding fields.
-	// This is the case for fields on types other than Query and Mutation.
-	d := f.Directives.ForName("hm_function")
-	if d != nil {
-		a := d.Arguments.ForName("args")
-		if a != nil && a.Value != nil {
-			v, err := a.Value.Value(nil)
-			if err == nil {
-				var list ast.ArgumentDefinitionList
-				var argName string
-				for _, val := range v.([]any) {
-					argName = val.(string)
-					fld := schema.ObjectDef.Fields.ForName(argName)
-					if fld == nil {
-						log.Printf("Field %s.%s does not exist", schema.ObjectDef.Name, argName)
-						continue
-					}
-					arg := ast.ArgumentDefinition{
-						Name: argName,
-						Type: fld.Type,
-					}
-					list = append(list, &arg)
-				}
-
-				return list
-			}
-		}
-	}
-
-	// No @hm_function directive, or no args argument.
-	// Just use the arguments on the field.
-	// This is the case for Query and Mutation fields.
-	return f.Arguments
-}
-
-func GetFunctionSchema(schema string) ([]functionSchema, error) {
-
-	// Parse the schema
-	doc, parseErr := parser.ParseSchemas(validator.Prelude, &ast.Source{Input: schema})
-	if parseErr != nil {
-		return nil, fmt.Errorf("failed to parse GraphQL schema: %+v", parseErr)
-	}
-
-	// Find all fields with the @hm_function directive and add their schema info
-	// to the map, using the resolver as a key.
-	var results []functionSchema
-	for _, def := range doc.Definitions {
-		if def.Kind == ast.Object {
-			for _, field := range def.Fields {
-				if field.Directives.ForName("hm_function") != nil {
-					schema := functionSchema{def, field}
-					results = append(results, schema)
-				}
-			}
-		}
-	}
-
-	return results, nil
 }
 
 type ModelSpec struct {
